+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Add `TextShow` instance for `CompactionFailed` in `Control.Exception`
 * Remove most monomorphic `TextShow` functions, as their utility is questionable, and their maintenance burden is undeniable
 * Rename `showbVersionConcrete` to `showbVersion` in `TextShow.Data.Version`
+* Add `SPECIALIZE` pragmas for `showbEFloat`, `showbFFloat`, and `showbGFloat` in `TextShow.Data.Floating`, mirroring the ones in `Numeric`
 * Refactor test suite
 
 #### 3.4.1.1 [2016.11.07]